requirements.txt
certifi==2021.10.8
click==8.0.3
cloudinary==1.28.0
colorama==0.4.4
Flask==2.0.2
Flask-MySQL==1.5.2
Flask-WTF==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyMySQL==1.0.2
python-dotenv==0.19.2
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.0
